fix: fixed ula signals and operations; add equal to ULAOut; makes simulator safer with 8 bit integers (but it is still under dev); improve reset functions and make this follow project pattern; make resets before the load instruction and load data

+++ src/encoding.c
@@ -6,7 +6,7 @@
 #include "utils.h"
 
 
-unsigned int binaryToUnsignedInt(const char *serializedBinary) {
+uint8_t binaryToUnsignedInt(const char *serializedBinary) {
     unsigned int decimal = 0;
     const size_t tamanho = strlen(serializedBinary);
 
@@ -34,13 +34,14 @@ void extendSignal(char *baseSerializedBinary, char *buffer, const int size) {
     strcat(buffer, baseSerializedBinary);
 }
 
-int complementOfTwoToInt(const char *serializedBinary) {
+int8_t complementOfTwoToInt(const char *serializedBinary) {
     const size_t size = strlen(serializedBinary);
 
     const char firstCharacter = serializedBinary[0];
     switch (firstCharacter) {
         case '0': {
-            return (int) binaryToUnsignedInt(serializedBinary);
+            // This number already is an 8-bit integer, the cast is only to convert a unsigned int on signed int
+            return (int8_t) binaryToUnsignedInt(serializedBinary);
         }
         case '1': {
             char binary[size + 1];
@@ -50,8 +51,8 @@ int complementOfTwoToInt(const char *serializedBinary) {
             char invertedBinary[size + 1];
             invertBinary(binary, invertedBinary);
 
-            const unsigned int value = binaryToUnsignedInt(invertedBinary);
-            return (-1) * (int) value;
+            const uint8_t value = binaryToUnsignedInt(invertedBinary);
+            return (int8_t) (-1 * value);
         }
 
         default: {
